refactor: remove add service page
